Abandon base64 in favor of base32 hex
We had some issue with base64 ids when transported between different
system we may not control, that would either change the case or try to
encode the "_" and/or "-" non alpha num chars. We also experienced
issues with some of our serialization that would concat an id with other
information using either "_" or "-" and would then result is errornous
decoding.

We concluded that it would be safer to go with base32 in exchange of
4 chars. We use the base32 hex charset to keep sortability.